+++ packages/workers/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 1.2.33
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/messaging@2.7.6
+  - @volontariapp/database@3.3.7
+  - @volontariapp/outbox@0.9.25
+
+## 1.2.32
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/messaging@2.7.5
+  - @volontariapp/database@3.3.6
+  - @volontariapp/outbox@0.9.24
+
 ## 1.2.31
 
 ### Patch Changes